Add postprocessing module for calculating stock grades and thresholds
